// src/product/enums/igv-affectation-type-id.enum.ts
export enum IgvAffectationTypeId {
  GRAVADO_ONEROSA = '10',          // Gravado – Operación Onerosa
  GRAVADO_RETIRO = '11',           // Gravado – Retiro por premio
  GRAVADO_RETIRO_DONACION = '12',  // Gravado – Retiro por donación
  GRAVADO_RETIRO_PROPIEDAD = '13', // Gravado – Retiro a favor de trabajadores
  GRAVADO_IVAP = '14',             // Gravado – IVAP
  EXONERADO_ONEROSA = '20',        // Exonerado – Operación Onerosa
  EXONERADO_TRANSFERENCIA = '21',  // Exonerado – Transferencia gratuita
  INAFECTO_ONEROSA = '30',         // Inafecto – Operación Onerosa
  INAFECTO_RETIRO = '31',          // Inafecto – Retiro por bonificación
  INAFECTO_RETIRO2 = '32',         // Inafecto – Retiro
  INAFECTO_RETIRO3 = '33',         // Inafecto – Retiro por muestras
  INAFECTO_RETIRO4 = '34',         // Inafecto – Retiro por convenio colectivo
  INAFECTO_RETIRO5 = '35',         // Inafecto – Retiro por premio
  INAFECTO_RETIRO6 = '36',         // Inafecto – Retiro por publicidad
  EXPORTACION = '40',              // Exportación de bienes o servicios
}